test(services): address PR #422 review feedback
Addresses Copilot + CodeRabbit review comments on PR #422:

**Sentinel collisions** (Copilot, character.rs line 476)
`0x7000_1100` already reserved by `player_load/core.rs:316`;
`0x7000_1200` already reserved by `vendor/purchase_helpers.rs:447`.
Stepped delete tests to `0x7000_1700`, visuals tests to `0x7000_1800`
(next free windows past the `system_options.rs` slot at `0x7000_1600`).
Updated doc-comments to document the new neighbour map.

**File cap split** (Copilot + CodeRabbit, character.rs lines 451-982)
`character.rs` was 983 lines, over the 700-line hard cap. Converted
to a directory module:
- `crates/services/src/base/character/mod.rs` (454 lines) — production
  + existing small inline test modules.
- `crates/services/src/base/character/delete_live_db_tests.rs` (218 lines)
- `crates/services/src/base/character/request_visuals_live_db_tests.rs`
  (390 lines)

**Defensive `if let Some` → `.expect(...)`** (CodeRabbit, world/tests.rs)
Replaced all 4 `if let Some(...)` blocks after `create_entity().unwrap()`
with `.expect(...)` so fixture invariants fail fast (per the project's
no-defensive-code-for-impossible-scenarios rule).

**`TryRecvError::Empty` not `is_err()`** (CodeRabbit, world/tests.rs)
Tightened both `rx.try_recv().is_err()` assertions to `matches!(...,
Err(TryRecvError::Empty))` so a `Disconnected` failure mode doesn't
silently pass.

**Unknown-op broadcast bytes** (CodeRabbit, world/tests.rs)
Strengthened `set_interaction_type_unknown_op_warns_and_does_not_mutate_flags`
to ALSO assert the broadcast fires with exact `0x42` LE u64 payload
and the channel ends empty. Documents the current behaviour (warn
falls through to broadcast) so a future change has to be conscious.

**Witness-broadcast byte-exact payloads** (CodeRabbit, witness_broadcast.rs)
Strengthened both `witness_entity_method_routes_one_packet_to_witness_addr_only`
and `entity_invisible_routes_one_packet_to_witness_addr_only` to also
reconstruct the expected wire bytes via `build_entity_method_packet` /
`build_entity_invisible` and assert byte-for-byte equality. Catches
wrong-builder / wrong-idbase / swapped-id-in-body / args-framing
regressions that the cardinality-only assertions missed.

**Visuals happy-path payload check** (CodeRabbit, character.rs lines 827-895)
Strengthened `handle_request_character_visuals_happy_path_sends_one_packet`
to capture the actual packet via `TestTransport::filter_to` and
reconstruct the expected bytes via `build_character_visuals`. Accepts
both possible orderings of the inventory-derived visuals (the SQL
join has no `ORDER BY`, so Postgres can return either permutation).
Catches wrong-bodyset, missing-visual, wrong-account_entity_id,
wrong-skin_tint, or wrong-message-type regressions.

Verified: cargo check + fmt + clippy `-D warnings` clean; 1109/1109
services-lib tests pass (no regressions vs PR base).
